backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
sqlalchemy==2.0.29
psycopg2-binary==2.9.9
alembic==1.13.1
pydantic==2.7.1
python-jose[cryptography]==3.3.0
httpx==0.27.0
PyJWT==2.8.0
python-multipart==0.0.9
